% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clarifai.R
\docType{package}
\name{clarifai-package}
\alias{clarifai}
\alias{clarifai-package}
\title{clarifai: Get Description of (Moving) Images}
\description{
Get Description of Images. Understand what is in the images. See \url{http://clarifai.com/}. 
Details about results of calls to the API can be found at \url{https://api.clarifai.com}.

Your need credentials to use this application. 
If you haven't already, you can get this at \url{https://developer.clarifai.com/}.
}
\author{
Gaurav Sood
}

