% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clarifai.R
\name{clarifai_check_auth}
\alias{clarifai_check_auth}
\title{Check if authentication information is in the environment}
\usage{
clarifai_check_auth()
}
\description{
Check if authentication information is in the environment
}

