% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clarifai.R
\name{clarifai_check_results}
\alias{clarifai_check_results}
\title{Check results}
\usage{
clarifai_check_results(res)
}
\arguments{
\item{res}{results being returned by the API}
}
\description{
Check results
}
