% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clarifai.R
\name{clarifai_CHECKAUTH}
\alias{clarifai_CHECKAUTH}
\title{Check if authentication information is in the environment}
\usage{
clarifai_CHECKAUTH()
}
\description{
Check if authentication information is in the environment
}

