% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clarifai.R
\name{clarifai_GET}
\alias{clarifai_GET}
\title{GET}
\usage{
clarifai_GET(path, query, ...)
}
\arguments{
\item{path}{path to specific API request URL}

\item{query}{query list}

\item{\dots}{Additional arguments passed to \code{\link[curl]{curl_fetch_memory}}.}
}
\value{
list
}
\description{
GET
}

