% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clarifai.R
\name{clarifai_POST}
\alias{clarifai_POST}
\title{Base POST AND GET functions. Not exported.}
\usage{
clarifai_POST(path, query = NULL, ...)
}
\arguments{
\item{path}{path to specific API request URL}

\item{query}{query list}

\item{\dots}{Additional arguments passed to \code{\link[curl]{curl_fetch_memory}}.}
}
\value{
list
}
\description{
POST
}
