% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/feedback.R
\name{feedback}
\alias{feedback}
\title{Provide Feedback on Tags of an Image}
\usage{
feedback(file_path = path_to_image, feedback_type = "add_tags",
  feedback_value = "")
}
\arguments{
\item{feedback_type}{must be one of the following types:
add_tags vector of suggested tags; default
remove_tags vector of tags you suggest be removed
search_click which tag(s) where clicked on
similar_docids similar documents
dissimilar_docids dissimilar documents}
}
\description{
Suggest tags, suggest removal of tags, suggest (dis)similar documents, or let clarifai know which tag was clicked on
}
\examples{
\dontrun{
feedback(file_path="path_to_image", feedback_type="add_tags", feedback_value="suggested_tag")
}
}
\references{
\url{https://developer.clarifai.com/}
}

