% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_info.R
\name{get_info}
\alias{get_info}
\title{Get Information}
\usage{
get_info()
}
\description{
Learn about the application -- what are maximum sizes allowed etc.
}
\examples{
\dontrun{
get_info()
}
}
\references{
\url{https://developer.clarifai.com/}
}

