% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_langs.R
\name{get_langs}
\alias{get_langs}
\title{List all the languages that the tag API call supports}
\usage{
get_langs(...)
}
\arguments{
\item{\dots}{Additional arguments passed to \code{\link{clarifai_GET}}.}
}
\value{
character vector listing all the languages that tag API call supports. See \code{\link{tag_images}} and code{\link{tag_image_urls}}
}
\description{
List all the languages that the tag API call supports
}
\examples{
\dontrun{
get_langs()
}
}
\references{
\url{https://developer.clarifai.com/}
}

