% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_token.R, R/tag_images.R
\name{get_token}
\alias{get_token}
\title{Get Access Token}
\usage{
get_token()

get_token()
}
\value{
a list with 4 things: access_token, token_type, expires_in, scope
}
\description{
Once you have set your Application ID and Secret via \link{\code{secret_id}}, get the access token.

Once you have your Application ID and Secret and have set it via \link{\code{setapp}}, get access token.
}
\examples{
\dontrun{
get_token()
}
\dontrun{
get_token()
}
}
\references{
\url{https://developer.clarifai.com/}

\url{https://developer.clarifai.com/}
}

