% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_usage.R
\name{get_usage}
\alias{get_usage}
\title{Get API usage for the current month and hour}
\usage{
get_usage()
}
\value{
named list
}
\description{
Get API usage for the current month and hour
}
\examples{
\dontrun{
get_usage()
}
}
\references{
\url{https://developer.clarifai.com/}
}

