% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_langs.R
\name{list_langs}
\alias{list_langs}
\title{List all the languages that the tag API call supports}
\usage{
list_langs()
}
\value{
character vector listing all the languages that tag API call supports. See \code{\link{tag_images}} and code{\link{tag_image_urls}}
}
\description{
List all the languages that the tag API call supports
}
\examples{
\dontrun{
list_langs()
}
}
\references{
\url{https://developer.clarifai.com/}
}

