% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tag_image.R
\name{tag_image}
\alias{tag_image}
\title{Tag Local Images}
\usage{
tag_image(file_path = NULL)
}
\arguments{
\item{file_path}{path to image file}
}
\description{
Tag Local Images
}
\examples{
\dontrun{
tag_image(file_path="path_to_image")
}
}
\references{
\url{https://developer.clarifai.com/}
}

