% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usage.R
\name{usage}
\alias{usage}
\title{Get API usage for the current month and hour}
\usage{
usage()
}
\value{
named list
}
\description{
Get API usage for the current month and hour
}
\examples{
\dontrun{
usage()
}
}
\references{
\url{https://developer.clarifai.com/}
}

