% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info.R
\name{get_info}
\alias{get_info}
\title{Get Information}
\usage{
get_info()
}
\value{
Named list of length 3: \code{status_code}, \code{status_msg}, and \code{results}. 
\code{results} is a named list of length 12. Contains information about max. and 
min. image and video size allowed etc. 

Prints \code{status_msg} by default
}
\description{
Basic information about the application --- what are maximum sizes allowed, 
default language, max. and min. image and video size, max. batch size, etc.
}
\examples{
\dontrun{
get_info()
}
}
\references{
\url{https://developer.clarifai.com/}
}

