% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/secret_id.R
\name{secret_id}
\alias{secret_id}
\title{Sets Application ID and Secret}
\usage{
secret_id(appdetails = NULL)
}
\arguments{
\item{appdetails}{Required; A vector of client_id, client_secret. Get these from \url{https://developer.clarifai.com/}.
Set them before you use other functions.}
}
\description{
Set Client ID and Secret. Needed for interfacing with Clarifai. Run this before anything else.
}
\examples{
\dontrun{
setapp(c("client_id", "client_secret"))
}
}
\references{
\url{https://developer.clarifai.com/}
}
\keyword{Client}
\keyword{ID}
\keyword{Secret}
\keyword{Sets}
\keyword{and}

