% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tag_image_url.R
\name{tag_image_url}
\alias{tag_image_url}
\title{Tag Remote Images}
\usage{
tag_image_url(img_url = NULL)
}
\arguments{
\item{img_url}{path to the image}
}
\description{
Tag Remote Images
}
\examples{
\dontrun{
get_image_urls()
}
}
\references{
\url{https://developer.clarifai.com/}
}

